[#89] "Intersection and contains" functionality
**New features and improvements**
- Added getIntersectionPoints() functionality which get two geometries and a list of points and returns the sublist of points within the intersection of the two geometries
- Added unit tests

**Related issue**
#89
